+++ backend/src/main/java/com/bdsa/disertatie/backend/service/FilmService.java
@@ -13,6 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -71,7 +72,7 @@ public class FilmService {
     }
 
     @Transactional
-    public void acordaNotaFilm (Long userId, String codFilmWikiData, Integer nota){
+    public void acordaNotaFilm (Long userId, String codFilmWikiData, BigDecimal nota){
         Film film = getFilmDupaUserIdSiCodWikiData(userId, codFilmWikiData);
         film.setNotaFilm(nota);
         filmRepository.save(film);
